Upgraded user page and fixed checkbox error in item class

+++ admin/users.php
@@ -5,7 +5,13 @@ require_once("globals.php");
 $title = USERS;
 $breadcrumbs = '<a href="index.php">' .ADMINISTRATION.'</a> > '.USERS;
 
-$query = "SELECT * FROM " . $db->quoteIdentifier('anyInventory_users') . " ORDER BY " . $db->quoteIdentifier('username') . " ASC";
+if ($admin_user->usertype == 'Administrator'){
+	$query = "SELECT * FROM " . $db->quoteIdentifier('anyInventory_users') . " ORDER BY " . $db->quoteIdentifier('username') . " ASC";
+}
+else{
+	$query = "SELECT * FROM " . $db->quoteIdentifier('anyInventory_users') . " WHERE " . $db->quoteIdentifier('id') . " = '".$_SESSION["user"]["id"]."'";
+}
+
 $result = $db->query($query);
 if(DB::isError($result)) die($result->getMessage().'<br /><br />'.SUBMIT_REPORT . '<br /><br />'. $query);
 
@@ -18,12 +24,10 @@ if ($result->numRows() > 0){
 				<td align="center" style="width: 15ex; white-space: nowrap;">
 					<nobr>';
 		
-		if (($admin_user->usertype == 'Administrator') || ($_SESSION["user"]["id"] == $row["id"])){
-			$table_rows .= ' [<a href="edit_user.php?id='.$row["id"].'">'.EDIT_LINK.'</a>] ';
-			
-			if ($row["id"] != ADMIN_USER_ID){
-				$table_rows .= ' [<a href="delete_user.php?id='.$row["id"].'">'.DELETE_LINK.'</a>] ';
-			}
+		$table_rows .= ' [<a href="edit_user.php?id='.$row["id"].'">'.EDIT_LINK.'</a>] ';
+		
+		if (($row["id"] != ADMIN_USER_ID) && ($row["id"] != $_SESSION["user"]["id"])){
+			$table_rows .= ' [<a href="delete_user.php?id='.$row["id"].'">'.DELETE_LINK.'</a>] ';
 		}
 		
 		$table_rows .= '
@@ -48,9 +52,13 @@ $output .= '
 			</td>
 		</tr>
 		<tr>
-			<td class="tableData" colspan="2">
-				<p style="padding: 5px;"><a href="add_user.php">'.ADD_USER.'</a></p>
-				'.$table_rows.'
+			<td class="tableData" colspan="2">';
+
+if ($_SESSION["user"]["usertype"] == 'Administrator'){
+	$output .= '<p style="padding: 5px;"><a href="add_user.php">'.ADD_USER.'</a></p>';
+}
+
+$output .= $table_rows.'
 			</td>
 		</tr>
 	</table>';
